Testsuite - fix federation tests expecting node count from one cluster
d3d13c6699a changed the tests to use get_nodes_by_state() to get
node_count from the system. But when a cluster is part of a federation
it's returning all nodes from the federation and the tests were
expecting just the number of nodes from the local cluster.

Bug 11420