Testsuite - Verify invalid LONG_OPT_SWITCHES is rejected
Test was checking that the opt.wait4switch was getting set to 0 when a
null data is handed to slurm_process_option_data() for
LONG_OPT_SWITCHES. A job can not run with zero switches and this option
should have have always been getting rejected as invalid. It was being
used a verification of the integer conversion code, so the implied
logical issues of having a job without a switch was missed as they are
caught by other code.

Bug 17182