Add NEWS entry
Note that invalid formatting of numeric data_t fields will now be
getting more strict checks and slightly modified conversion results for
edge conversion cases. All of this is a behavior change requiring a NEWS
entry even though I consider the prior behavior to be a bug or atleast
non-intuitive.

openapi/v0.0.38 is explicitly notated in the NEWS entries as it is the
last user of the unversioned data_t parsing functionality in slurm_opt
handling code. This is where the majority of the explict conversion
request to DATA_TYPE_INT are being called with considerably less care to
handle meta data semantics of non-integer values, thus increasing the
chances a user will get a unexpected and different result.

The new conversion behavior isn't a large change from <=23.02 behavior
where regular expressions were used to validate all numeric values but
miss invalid formatting where sscanf(%u) or atoi() will still parse out
a valid integer or float.

There are many locations in the parsing code where there is an explicit
call to data_convert_int(DATA_TYPE_INT) which activates the forced
conversion mode. The forced conversion mode will not result
in a conversion that loses accuracy but may result in loss of precision
for operations. An common example is where a floating point number is
converted to an integer.

Conversion of null to numeric zero in forced conversion mode will also
now potentially happen in more locations where integers are expected.
Prior behavior would have resulting in a rejected conversion of null ->
integer.

Bug 17182, 15042