sacct - show 0 AllocCPUS for pending and canceled jobs with no allocation
print_fields() was taking the allocated cpu count from a job's req tres
field when the alloc tres one was empty. This was causing sacct to display
a misleading number (not 0) for AllocCPUS for a pending job or one that was
canceled before it had an allocation. Make it so this function doesn't do
that and correctly show 0 for AllocCPUS for such jobs.

Bug 17092