Fix --gpu-bind=single binding tasks to incorrect gpus
The logic for determining the best gres to the task happens in each task.
Each task calculates the best gres for all tasks leading up to itself.
Therefore, the logic is supposed to be identical for each task.
The bug is that the cpu_set of the current task was being used to calculate
the best gres for every task.

For example, when task 2 calculated gres for task 0, it found the gres
that was best for task 2 and removed that gres from gres_slots. So by the
time task 2 calculated the gres for itself, the best gres was already
removed from gres_slots, so it selected a suboptimal gres.

Bug 17113