Notify clusters in federation when cluster registers
When clusters in a federation register all at the same time, it's
possible that the sibling's cluster information might not have their
respective host and port filled in which would leave the cluster's
unable to know how to communicate with the siblings.

This situation can be seen easier by using CommitDelay=#.

When a cluster registers and updates its host, and rpc we should let the
other clusters in the federation know.

Bug 17061