Move a UID validation example from cli_filter to job_submit
In Slurm 23.02, options['uid'] in cli_filter.lua does not return the UID of
the user but returns the result of --uid. This change was made to allow
Linux user namespaces to work. In addition, because a cli_filter is not
secure, an example of rejecting a job due to an invalid constraint is better
done in a job_submit plugin.

For these reasons, move and adapt the example code checking UID and
interactive jobs to job_submit.lua.example.

Bug 17075