Fix gres_select_filter_sock_core for CR_CPU_MEMORY
In case of CR_CPU_MEMORY can_job_run_on_node can set number of CPUs
available to the job based on --mem-per-cpu to a value that is not a
multiple of ThreadsPerCore. In this case we can't assume that removal
of core always removes ThreadsPerCore. Instead of that assumption just
reset the value number of available cores * ThreadsPerCore.

Bug 17229