Default to one cpu per task if cpus_per_gres requests cpus < tasks
A job or step may request cpus_per_gres such that the cpu count could be
less than the number of tasks on a node. For job allocations, we assume
one cpu per task if that happens. Do the same thing with step allocations
instead of rejecting the step.

Bug 16809