Send slurmctld timeout to slurmstepd
If using a backup slurmctld and the primary goes down temporarily,
slurmstepd uses the slurmctld timeout to determine when to retry
contacting the primary. Send the value to slurmstepd so it waits the
correct amount of time, rather than NO_VAL16/2.

The slurmctld timeout is initialized to NO_VAL16 in init_slurm_conf()
and the sleep time of slurmctld_timeout / 2 happens in
slurm_send_recv_controller_msg().

Bug 17232