Added error message when using sattach on batch or extern steps.
Before sattach $JOBID.batch would output nothing and end when the batch
step ends. This is due to sbatch shortcutting its i/o straight to a file.
Now when a batch step is specified the error "Cannot be used with extern
or batch steps" is outputted. The error message for attachng to extern was
also updated to this.

Bug 17200