Revert "Make it so srun no longer inherits SLURM_CPUS_PER_TASK"
This reverts commit d848512f4050a4692a331a50c79b56831bff94c6.
That commit was done because in slurm <= 21.08 srun would set
--cpus-per-task when reading SLURM_CPUS_PER_TASK, for example from within
an allocation. This didn't work unless you set --exact, so we needed to
imply --exact when using --cpus-per-task. Nevertheless, this broke mpirun
calls when for example using '-c1'. The spawned 'sruns' would be bount to
only one cpu. So there was an incompatibility between mpirun launchers and
the requested resources.

With the introduction of srun --external-launcher, now mpirun will get all
the resources of a node and will ignore the --cpus-per-task parameters, thus
effectively getting bound to all the allocated cpus. Now we want to recover
the ability to have this variable set and read by srun in normal allocations,
so we revert back this change without any affectation to mpirun.

Bug 15831