Fix userspace governor with frequency ranges
A request with --cpu-freq having p3 as userspace and a frequency
range set in p1-p2 for min and max doesn't make sense, since UserSpace will
set a static frequency.

The documented behavior was not working either, since theorically the new
frequency should be set to the closest limit if it was out of range, and
that was not happening.

Fix the behavior by not accepting a range when using Userspace governor.

Bug 16783