Fix cancelling pending federated jobs from non-origin cluster
The sibling cluster would just handle the cancel and let the origin
cluster know that the job was cancelled, but wouldn't revoke any
remaining sibling jobs.

This is a partial revert of commit 8ae0332b09.

For pending jobs, route the cancel back to the origin cluster to remove
all active pending sibling jobs since it knows where they reside.

Bug 17285

Signed-Off-By: Ben Glines <ben.glines@schedmd.com>