Fix how arbitrary distribution allocates cpus.
Before, the distribution of cpus did not depend on the number of tasks
the arbitrary distribution put on each node. This would lead to nodes
with multiple tasks not having enough cpus allocated to them even though
the job as a whole had the right number of cpus. Now the
select/cons_tres plugin allocates cpus to nodes based of the number of
tasks each node will run. The req_nodes is used to calculate the number
of tasks on each node so for arbitrary distributions the nodelist is
left with duplicates.

Bug 17039