cons_common - make step 2 changes to node_map permanent
All the resources marked not available to the job in _job_test Step 2
are not available to the job in any case.

The specific case here is that our job is exclusive and there are
resources on a node that we can neither gang-schedule nor preempt.
In this case we remove the node from node_bitmap in _block_whole_node.
If we don't make this permanent we won't remove the node in next step
since we don't check for non-preemptable jobs there.

Bug 17237