In case of CR_SOCKET, don't clear cores on sockets in use.
The logic in gres_select_filter_sock_core() attempts attempts to remove
cores the job doesn't require. For CR_SOCKET the whole socket is
allocated to the job so we should just leave the whole socket allocated.

At the end of gres_select_filter_sock_core(), we have a loop that spreads
the job on "required sockets". If selected GRES isn't bound to socket
all sockets are treated equally, which results in the job being spread
across sockets. If full sockets are allocated (CR_SOCKET) than we end up
allocating more than was required.

Bug 17288