Fix how the nodelist from SLURM_HOSTFILE is shortened by -n option
Before the number of hosts in the nodelist would only be correctly
shortened if the SLURM_HOSTFILE only contained one host per line.
However, the file can have comma separated hosts and repeat hosts with
the * notation (ex: n1*5). This is included in the fix for arbitrary
distribution because it specifically mentions that the SLURM_HOSTFILE
can contain comma separated hosts.

Bug 17039