node_features/helpers - fix logging an error on success
node_features_p_job_xlate() figures out which node features are valid for
the job, and returns only the list of changeable features. If the picked
features are all static features, then this will return NULL. This is still
considered a success; do not log an error in this case.

Bug 17111