data_parser/v0.0.40 - Add single flag marker for parsers
In many instances, a flag array can only represented as a single flag
which negates the need to define a field as an array. Instead of adding
wrappers to parse and then extract the first array index, this just
skips the process and only has a single string to represent the flag
array in OpenAPI format and in the output.

Bug 17457