Remove MYSQL_OPT_RECONNECT completely
This shouldn't affect anything as we are forcing it to be off in the
current code and this just removes that. The default is off.

Beginning with MySQL 8.0.34, the automatic reconnection feature is
deprecated. The related MYSQL_OPT_RECONNECT option is still available
but now returns a deprecation warning to the standard error output if
your application calls the mysql_get_option() or mysql_options()
function with the option.

Bug 17477