NEWS for the prior four commits
Set SLURM_NTASKS when --ntasks-per-node is requested

The prior commits guarantee that SLURM_NTASKS is set in the job
environment when the option --ntasks-per-node is requested. This
effectively reverts the intention of commit ef513023ad to only set
SLURM_NTASKS when --ntasks is explicitly requested.

The change in 23.02 to not set SLURM_NTASKS when --ntasks-per-node is
requested was made late and without any documentation in RELEASE_NOTES
NEWS, or any man pages. Apparently lots of people rely on this, so we are
reverting the change in 23.02.

Bug 17108