data_parser/v0.0.40 - Use correct pointer in _foreach_dump_list()
dump() requires a pointer to the parser object. The addpl() macro
definitions had previously effectively worked around this by having the
parser type that corresponded to the pointer type which was hidden by
secondary calls to dump which grabbed the pointer of from the function
on the stack. While this worked, it breaks completely when trying to
dump a list of strings (char *) as the dump function will get the wrong
pointer (char **) and de-reference it. This wasn't an issue until
v0.0.40 when new parsers triggered this code path.

Bug 17530