gpu/nvml - Get other gpu statistics even if one fails
If nvml functions fail for a single device, we should still continue to
query all devices. We're doing a best effort in trying to get the memory
and gpu utilization, and continuing will give us a better chance to get
all the stats.

Similarly, if _get_gpumem() fails, we should still try to do _get_gpuutil().

Bug 17296