slurmdbd - avoid potential NULL pointer dereference in mysql plugin.
If the only account associated with a user is missing, the parent_id may
be NULL and cause a NULL pointer dereference.

This patch catches this scenerio and treats it as if the parent account
has no limits to allow both the slurmdbd and slurmctld to continue to
function.

Bug 16338