data_parser/v0.0.39 - fix updating qos for associations
Originally any request for POST /slurmdbd/v.0.0.39/associations that
contained the "qos" field would fail because the list of available qos
was not being initialized before calling resolve_qos() which requires
it. This was because the parser of QOS_STRING_ID_LIST was marked as
NEED_NONE even though it does need qos. This caused
load_prereqs_funcname() to not load in the qos list.

Bug 17476