slurmdbd - Avoid printing error when cluster registers
The slurmdbd attempts to send db updates to all registered clusters
after committing. With changes in 8aa6000e2c, the db tries to send the
updates to the freshly registered cluster before the cluster has started
its rpc manager thus resulting in error messages when failing to connect
to the cluster.

We avoid this by not sending updates to the newly registered cluster.
The error messages could still happen if updates happen after
registering and before the rpc manager is up -- which is a small window.

Bug 17550

Signed-off-by: Danny Auble <da@schedmd.com>