Use gethostname() for x11_alloc_host if IP is localhost.
If resp_host is 127.0.0.0/8 or ::1, the srun/salloc was most likely
executed from the slurmctld node. In this case its would be better to
use gethostname() to have the slurmd do the lookup.

This approach fixes a regression in 23.02.3 where a login node submits the
job, but its MUNGE credential was encoded with a loopback address in the
host field. (This usually happens when the hostname is mapped to the
loopback address in /etc/hosts.)

Bug 17491.