Avoid killing extern step on all nodes when using --no-kill
Without this patch, when a node is being removed from an allocation that
uses --no-kill the extern step gets killed on all nodes which broke things
on the remaining nodes that rely on the extern step such as x11,
job_container/tmpfs, and pam_slurm_adopt.

To avoid this, when one or more nodes fail, only kill the extern step on the
failed nodes. The extern step will still be killed when the job completes or
is killed. If the job did not specify --no-kill and a node failes, then the
job will be killed anyway including the extern step.

Bug 17395