Testsuite - Improve test32.8 verifying fanout behavior
Although previous implementation was using the --fanout option it was
not verifying its behavior, but the exact same behavior that was already
verified in test32.4.
New implementation uses sacct to check the expected number of sruns
launched by sgather. To double check the behavior, test32.4 also adds
the same check (with different expected results).

Bug 17482