set and reset TRES job_desc fields for validating GRES in a job update
In case of update of GRES specification we call gres_job_state_validate
on the job_desc requested by the update. If some member of the struct
wasn't specified we need to use what we had before so we need to set the
TRES job_desc fields. Reset them after validating.

This is preparation for the next commits and implicitly fixes CpusPerTres
not being updateable.

Before:
$scontrol update job=55 CpusPerTres=gres:gpu:2
Invalid generic resource (gres) specification for job 55

After:
$scontrol show job | grep CpusPerTres
   CpusPerTres=gres:gpu:1
$scontrol update job=55 CpusPerTres=gres:gpu:2
$scontrol show job | grep CpusPerTres
   CpusPerTres=gres:gpu:2

Bug 17184