Improve handling of slurmd shutdown on scontrol shutdown
Instead of sleeping for 10ms as a "should be enough" to communicate with
slurmd's wait for the dedicated agent thread to terminate.

In this case we skip _notify_slurmctld_nodes. We're shutting down so
registering non-responsive nodes isn't that important and the function
requires write lock on nodes, while shutdown RPC already owns a read
lock.

Bug 17611