Allow jobs that request newly configured gres types to be queued.
Previously if the slurmctld restarted with a new gres type configured on a
node no jobs could be queued that used that gres type until the node
registered. This is due to the topo values being saved from the old
configuration. Since they will be updated once the node registers default
to using the types instead of topo values.

Bug 17433