Use pw_name from the credential to set SLURM_JOB_USER in build_env
In the epilog, the SLURM_JOB_USER env var was relying on a lookup on the
node as job_env->user_name was never populated.

Instead, directly use the pw_name field from the credential. Only fall-back
to a local lookup if that was not provided. This will be set in the cred
for all systems that have not explicitly configured "disable_send_gids".

As this is now expected to work in most cases, remove the specialized
handling for Cray Aries systems from the fall-back path.

Bug 17094