Avoid using AI_PASSIVE in get_addr_info()
If glibc and the kernel get out of sync, then glibc will end up segfaulting
if the size of struct nlmsghdr has changed.

While the solution is to ensure that the running glibc is compatible with
the kernel, this is not a tenable solution for sites using containers to run
Slurm. These containers can naturally fall out of sync with the kernel as
the host is updated but the containers are not (or vice versa). By not using
AI_PASSIVE, and instead setting ai_addr directly, then this issue can be
avoided.

Bug 17171