auth/jwt - Ignore JWKS x5c field
Revert commit 1572786459.

The 'x5c' field is a base64 wrapped DER encoded x509 certificate which
is not supported by libjwt currently, causing otherwise valid JWKS files
to result in JWT decode failures.

The original code expected a string instead of the array (per RFC7517),
which resulted in it being ignored (and the key being reconstructed
from 'e' and 'n' instead) in almost all cases when a valid JWKS was
presented, which is how this issue wasn't caught previously.

Bug 16586