Kill recovered jobs that require none configured gres types.
Previously, if a job that required a typed gres was running and then the
slurmctld restarted with a new gres type configured that replaced the
one the job was using, you would get a "This should never happen, we
couldn't find the gres" error. To prevent this, check to see if the
recovered job is using valid gres types.

Bug 17433