Stop tracking the siglen in the slurm_cred_t.
For both plugins the signature is plain-text anyways.

For the RPCs, packstr vs packmem are formatted identially, the difference
is only that packstr will call packmem with strlen + 1 automatically.

For cred/munge, the siglen is strlen + 1 already, so they're equivalent.
(And the siglen is already ignored for verifying the cred.)
For cred/none, the siglen is strlen - so switching does result in clients
seeing an extra byte on the unpack. However, the verification is just
as happy to see siglen + 1 - it'll end up comparing the NUL-terminating
byte from the reference string and the provided signature and passing.