gres - rebuild GRES core bitmap for nodes at startup
If a node changes its core count, for example in slurm.conf, and
slurmctld is restarted, existing jobs assigned to this node will keep
the core bitmap with the old size.

This commit fixes error: Core bitmaps size mismatch on node [HOSTNAME],
which causes jobs to enter state "Requested node configuration is not
available".

Bug 17651