Docs - Include note about length of UnkillableStepTimeout
In the case of an interactive job (srun or salloc) at the moment the
job is cancelled, a message is sent to the srun or salloc command in
order for it to close itself. If within 5 tries this message cannot be
delivered, the process is assumed to be non-responsive and the job
finishes normally. But if this period is smaller than
UnkillableStepTimeout, then the timer for that will hit first and drain
the node(s) that the job is being run on, assuming that there is
something wrong with those nodes and that the processes could not be
killed, when the issue is the srun or salloc not responding in time.

Bug 13419