Initialize max_node_cnt to NO_VAL
Initialize max_node_cnt to NO_VAL so that max_node_cnt=0 can be a valid
number.

On startup, max_node_cnt is NO_VAL which lets nodes from the slurm.conf
to be read in. After reading in nodes from slurm.conf, max_node_cnt is
set to the max of the read in records and the configured max_node_cnt.

Bug 17667

Signed-off-by: Chad Vizino <chad@schedmd.com>