Properly signal a non-meta array job task in job_str_signal()
If job_str_signal() is called with a stringified non-meta array task job
id (ex. "12345") then it will fail. This causes problems when using
scancel or in slurmd for an array job when a prolog or the batch job
fails to launch and the job must be killed or if failing to requeue (ex.
job did not permit requeues).

This patch applies similar code from update_job_str() for handling
single jobs or non-meta array tasks where numeric job ids have been
stringified.

The previous logic in job_str_signal() was not correctly identifying a
non-meta array task and so would fail when trying to process such a job
id as a meta array one.

Bug 17638