Signal only a meta array job when it fails to be launched by slurmd
When a meta array job (or prolog for one) fails to start in slurmd, an
attempt is made to requeue it and if that fails (ex. --no-requeue set on
job) then it is killed.

Previously, killing a meta array job caused the entire array job to be
killed. This fixes that by using KILL_ARRAY_TASK to indicate to
job_str_signal() that only the single task should be killed and not the
whole array job.

Bug 17638