Fix incorrect memory.swap.max setting in cgroup/v2
In cgroup/v2 there are two independent limits, memory.max and
memory.swap.max. In cgroup/v1 the swap limit was memory+swap, so it
needed to be set to memory.limit_in_bytes + a percentage, so
AllowedSwapSpace=5 implied setting memory.memsw_limit_in_bytes to 105%
of memory.limit_in_bytes.

In cgroup/v2 we need to separate these values, and just make memory.swap.max
to be a 5% of memory.max, not 105%.

Bug 17233