auth/jwt - fatal() if keys are insecure.
Print a warning() if they're not owner by SlurmUser/root, but don't
consider this an invalid configuration at this time.

The jwt_key should not be accessible by (other).

The JWKS should not be writable by (other), but there are no secrets there
so read/execute are fine if set.

Bug 17325

Co-authored-by: Megan Dahl <megan@schedmd.com>