Fix indexing for parent path in _foreach_flag_parser.
The index should not be based off of the number of times all of the flags
have been looped through. Only increment the index when it is done with
the current data_t list element.

Since the index is not updating as often there is no need to call
_flag_parent_path() more then once.

Bug 17782