data_parser/v0.0.39 - Fix parsing of flag arrays to set more then one.
Before when parsing any flag array only one flag would end up being set,
that being the last flag in the array. This is because _foreach_flag_parser
gets called for every flag in the array to be parsed. This is an issue
because _foreach_flag_parser() loops through every flag that it could be
and if it doesn't match the wanted flag it unsets it from result. This
means if you had the QOS flags ["NO_RESERVE", "NO_DECAY"] the first time
_foreach_flag_parser() is called it set "NO_RESERVE" in the result but the
second time through it "NO_RESERVE" is unset because it doesn't match
"NO_DECAY" and then it sets "NO_DECAY" in the result.

Now _foreach_flag_parser() only unsets any given bit once in the dst. Now
the first time through _foreach_flag_parser() will unset everything that is
needed and then the rest of the times it is called it will only set bits.

Bug 17782