Prune redundant logic and use SHOW_MIXED
NODE_STATE_MIXED is only used by the client for display purposes. When a
node data request is made, the SHOW_MIXED flag will have the API to
alter the response data to reflect MIXED nodes.

The clients we effectively doing the SHOW_MIXED logic themselves instead
of requesting it from the API.

Bug 17846