Export topology_g_build_config().
This needs to happen for the route/topology plugin.  It calls
topology_g_build_config() but if the exe is linked against libslurm.so
topology_g_build_config() isn't exported since commit 73d146253a
when we renamed it from slurm_topo_build_config() which was exported then.

The topology plugin is inited in slurm_send_recv_msgs() for non-daemons.