Only use the per-node cpus_per_task in step_layout if needed
Fix a regression in which cpu binding by the task plugin was inconsistent.
This regression was caused by the following commits:

5839f94a3e
d2bd871fbd
ad83659f70

These commits stored cpus_per_task in a per-node array that was then passed
to slurmd and to the task plugins. The values in the cpus_per_task array
were sometimes computed differently than the behavior prior to these
commits.

This commit fixes the inconsistency by only using the per-node cpus_per_task
array if cpus_per_tres (currently only --cpus-per-gpu) was requested. We
use the old behavior when not using --cpus-per-gpu.

Bug 17711