Set is_dynamic flag for FUTURE nodes
Allow clients to lookup node address for dynamic future nodes.

If a node is State=FUTURE in the slurm.conf, the node could be marked as
idle with scontrol, but the man page says to update the slurm.conf when
doing so. So the clients won't see the nodes as dynamic in this case.
If a client it talking to a State=FUTURE node then we can assume that
the node is a dynamic future node.

Bug 16752