Avoid backfill planning on completing and non-available nodes
Completing nodes are assumed to be returned to idle soon, but sometimes
nodes can stay in completing states for a longer time. Since backfill
assumes that nodes will return to service soon, it will plan jobs
on completing nodes, which can lead to higher priority jobs to getting
starved because they are waiting for completing nodes to finish but
could have run sooner on non-completing nodes.

This change gives completing nodes, and other states where the node will
not come up for some time, a higher weight to avoid planning jobs on
these nodes.

Bug 16741

Co-Authored-by: Dominik Bartkiewicz <bart@schedmd.com>