Fix issue when dealing with cross version slurmd's (some on 23.11 and some on 22.05)
In this case the 23.11 ctld would pack the lowest version (22.05) which
on the 23.11 nodes would try to free the select_jobinfo message that was
packed going to the 22.05.  Since we don't load the select plugin in the
slurmds anymore we would assert on the issue.

This moves the assert lower in the function to only matter when we have
->data, which in 23.11 we will not.

Continuation of 7a5d138e76