data_parser/v0.0.39 - Prevent xassert after DUMP_FUNC(QOS_ID) error
If the qos_id is not found in the qos list DUMP_FUNC(QOS_ID) returns
without setting a string in the data_t. This causes an xassert error during
_check_dump() since data_get_type(dst) is DATA_TYPE_NULL while
openapi_type_format_to_data_type(parser->obj_openapi) is a
DATA_TYPE_STRING.

Instead of leaving it as a DATA_TYPE_NULL set the string to be Unknown,
which is also what sacct's normal output shows for any bad QOS IDs.

Bug 17895