data_parser/v0.0.39 - Fix how jobs with deleted QOS are dumped
Before the list of QOS provided by NEED_QOS did not included deleted QOS.
This causes issues when dumping jobs that used old deleted QOS. In
DUMP_FUNC(QOS_ID) if the qos_id is not found in the QOS list provided by
NEED_QOS it return an error and stops the parsing of the jobs early.
To prevent this include the deleted qos in the QOS list.

Bug 17895