Avoid revalidating credentials.
MUNGE will, on non-developer builds, complain about replayed credentials
when sending multiple blocks for the same file, or for the library files.

The 'void *' argument is to avoid an issue where the type is defined in
slurm_protocol_defs.h, but cred.h is included at the top of that header,
leading to an awkward cyclical include path that's best avoided.

Regression from be34104add, but much easier to fix now.