acct_gather_energy/ipmi - Simplify initialization logic
_open_dcmi_context will exit early if the context is non-NULL.
The dcmi ctx is thread local, so a separate thread would need to open a
new dcmi context anyway.
_init_ipmi_config called _open_dcmi_context if _thread_init did not.
Therefore, we do not need to check if this is the first call before calling
_open_dcmi_context; instead we can always call _open_dcmi_context here.

This means that the call to _open_dcmi_context from _init_ipmi_config
is now redundant and is also removed in this commit.

Bug 18013