Improve srun messages when retrying step creation
This differenciates logged messages between the case where srun retries
step creation because of a timeout vs the situation when it retries step
creation on other step completion notification. If we retry notified by
slurmctld the info message about nodes being busy is missleading since
it's actually printed when resources got freed.

The mechanism used for notification is based on general trust, any
connection to the port is interpreted as notification about step
completing in a job.

Bug 16411