Fix fd socket name resolution
Partially revert 73476b0f6cd1d which was using realpath and being unable to
resolve socket symlinks. realpath() returns ENOENT over socket symlinks so
we want to use readlink. Use it as in the orginial behavior in eb0755b9c4b
but instead of getting the size with lstat/stat (which is not possible over
a symlink), use the PATH_MAX.

Bug 10214