Make TRES related env vars in a step override the env vars in the job
If a job requested a tres related option whose environment variable is set
by set_env_from_opts(), then srun requested that same option but with a
different value, the environment variable was not updated: it was still the
value from the job's environment, not the value requested by srun.

This commit makes options requested by srun override the environment
variable in the job's environment.

Bug 17610