Respect --cpus-per-gpu per node, not just per job
Previously, you could have a job allocation that honored the
cpus_per_gpu request as a whole, but not on each node. For example:

$ salloc --cpus-per-gpu=1 --gpus-per-node=3 --nodes=2

This job requests 6 total cpus. It previously could be allocated
4 cpus on one node and 2 cpus on the other node, with 3 gpus on each
node. This does not respect the --cpus-per-gpu request on each node.

Commit 191153390e tests cpus-per-gpu for step allocations. After this
commit, a step request in the previous job allocation was rejected due
to not enough cpus on each node:

$ srun sleep 1
srun: error: Unable to create step for job 576: More processors
requested than permitted

This commit honors --cpus-per-gpu on every node in the job allocation by
setting a maximum number of tasks per node, which is then used in task
distribution to select cpus on each node.

Bug 17609