sackd - Fully implement reconfiguration mechanism.
Reconfigure is now handled by fork()+exec()'ing a child process with
the updated configuration files.

If it can successfully start up from those, it will return SLURM_SUCCESS
over a pipe to the parent and the parent will exit.

If it fails to start, the parent process will resume execution.