Rebuild gres_node_state_t's topo if configs differ from state file.
Due to gres_node_state_t topo values being saved in the save state file in
commit a54f9cffff, if the number of gpus is changed in the config files
before a restart, those changes would not be reflected in the topo values if
the topo counts of the slurmd_conf_tot and gres_ns were the same.
E.g. in slurm.conf changing GRES=gpu:intel:2 to GRES=gpu:intel:3.
This would also not allow the types of gpus to change on restart either.

Now the topo values are rebuilt if the configs are updated and there is a
topo_cnt even if the old topo_cnt matches the new one. This does allow the
type to be updated during a scontrol reconfig leading to the need to remove
any types that have an type_cnt_avail[i] = 0 to be removed. This is because
if they are not removed then the topo values will be rebuilt every time
a node registers.

Bug 17433