data_parser/v0.0.39 - Fix how jobs with deleted associations are dumped
Before the list of associations provided by NEED_ASSOC did not included
deleted associations. This causes issues when dumping jobs that used old
deleted associations. In DUMP_FUNC(ASSOC_ID) if the association is not
found in the list provided by NEED_ASSOC it return an warning.
To prevent this include the deleted associations in the associations list.

Bug 17895