data_parser/v0.0.39 - Fix how jobs with invalid QOS IDs are dumped.
In DUMP_FUNC(QOS_ID) if the qos_id is not found in the QOS list provided by
NEED_QOS it return an error and stops the parsing of the jobs early.
Instead of returning an error it should return a warning so that the data
can be dumped gracefully without being cut short.

Bug 17895