Fix parsing log messages with %n
%n can be interpreted by vfprintf if passed inside a variable containing
the format string. In C printf() %n is a special format specifier which
instead of printing something causes printf() to load the variable pointed
by the corresponding argument with a value equal to the number of
characters that have been printed by printf() before the occurrence of %n.

This could lead to segfaults e.g. from pmix code when printing
_pmixp_job_info.server_addr_unfmt in combination with
LOG_FMT_FORMAT_STDERR.

Fix it by specifying the format and arguments directly in _log_msg.