data_parser/v0.0.40 - Fix how jobs with deleted QOS are dumped
Before the list of QOS provided by NEED_QOS did not included deleted QOS.
This causes issues when dumping jobs that used old deleted QOS. In
DUMP_FUNC(QOS_ID) if the qos_id is not found in the QOS list provided by
NEED_QOS it sends a warning which is not what users expect if the QOS is
only deleted. To prevent this include the deleted qos in the QOS list.

Bug 17895