cred/munge - Save signature in cred.
In the slurmctld this is unnecessary, which is why it was removed as
part of the broaded cred/munge refactor - the buffer is sufficient.

But for 'srun -Z', we still need the signature - in addition to the
buffer - to validate inbound I/O. Otherwise the step launch will
fail with "srun: error: Invalid IO init header signature".

(For normal launch, both signature and buffer are available in srun
since they'll have been populated by slurm_cred_unpack().)

Bug 17981.