Rework reconfigure mechanism again.
Use the main loop rather than working out a dedicated shutdown/startup
pattern to deal with failed reconfigurations.

This fixes a race condition that was exposed in the previous version,
where a simple "srun id; scontrol reconfigure" would leave a stray
job in place. Side-step working out the exact mechanism for that,
and instead unify the reconfiguration alongside the HA handoff code
here. We do need to avoid calling read_slurm_conf() - that's the
most likely cause of a failed reconfig - but we can avoid that with
some minor adjustments to check if this is a reconf or not.

This also fixes an issue seen in the testsuite where multiple reconfigs
were being attempted in rapid succession. Rather than return EINPROGRESS,
queue those up here and return the results to all of them.