Rework slurmdbd agent shutdown.
Avoid spamming the signal into the thread, just send it once and wait.

pthread_timedjoin_np() would be a nice alternative to the additional
conditional, but it's non-portable.

Add note that the pthread_cancel() call is problematic - this could
fire while, e.g., the agent thread was holding a glibc lock and cause
the process to crash.

Saves another 100ms on reconfig.