data_parser/v0.0.40 - Prevent DUMP_FUNC(QOS_NAME) from hitting xassert.
Previously, if the QOS name was NULL then data_set_string() would set the
data type as TYPE_NULL. This would cause the xassert(data_get_type(dst) ==
openapi_type_format_to_data_type(parser->obj_openapi)) in _check_dump() to
be hit. This is because it is expecting a string type not a null type.

Bug 18184