data_parser/v0.0.41 - Continue dumping even if accounting is disabled.
Merge commit fb35bd7c46 into v0.0.41.

Previously, if the accounting type is none then the args->db_conn will be
NULL. This would cause the _db_query_list() call to hit a xassert(). If
xasserts are removed it also leads to _db_query_list() to error() because
the list it retrieves will be NULL due to the connection being NULL. These
errors would cause the dumping command to end.

These changes allow for the dumping to continue with empty lists, but does
present the user with a warning that the needed data could not be
retrieved.

Bug 18184