data_parser/v0.0.40 - warn when attempting to dump an empty shares list
DUMP_FUNC(ASSOC_SHARES_OBJ_LIST) hits the xassert in list_for_each()
because resp->assoc_shares_list is NULL when the accounting storage type is
none. To prevent this return instead of calling list_for_each() if the list
is NULL.

Added warning that the shares list is empty because the storage type is
none.

Bug 18184