Fix oversubscription on partitions with PreemptMode=OFF
In "test 2" of _job_test we verify if we have any chance to start the
job. If the partition job is submitted to has multiple rows we may be
able to use the core in one of those even if the core is allocated in
the other. This usually happens in the first if statement if the
partition doesn't have PreemptMode=OFF set.

Bug 18019